@@ -20,13 +20,74 @@ interface IRequest {
 @injectable()
 class CreateOrderService {
   constructor(
+    @inject('OrdersRepository')
     private ordersRepository: IOrdersRepository,
+    @inject('ProductsRepository')
     private productsRepository: IProductsRepository,
+    @inject('CustomersRepository')
     private customersRepository: ICustomersRepository,
   ) {}
 
   public async execute({ customer_id, products }: IRequest): Promise<Order> {
-    // TODO
+    const customerExist = await this.customersRepository.findById(customer_id);
+
+    if (!customerExist) {
+      throw new AppError('Could not find any customer with the given id');
+    }
+
+    const existentProducts = await this.productsRepository.findAllById(
+      products,
+    );
+
+    if (!existentProducts.length) {
+      throw new AppError('Could not find any products with the given ids');
+    }
+
+    const existentProductsIds = existentProducts.map(product => product.id);
+    const checkInexistentProducts = products.filter(
+      product => !existentProductsIds.includes(product.id),
+    );
+
+    if (checkInexistentProducts.length) {
+      throw new AppError(
+        `Could not find any product${checkInexistentProducts[0].id}`,
+      );
+    }
+
+    const findProductsWithNonQuantityAvaliable = products.filter(
+      product =>
+        existentProducts.filter(p => p.id === product.id)[0].quantity <
+        product.quantity,
+    );
+
+    if (findProductsWithNonQuantityAvaliable.length) {
+      throw new AppError(
+        `Could not find any product ${findProductsWithNonQuantityAvaliable[0].quantity} is not avaliable ${findProductsWithNonQuantityAvaliable[0].id} `,
+      );
+    }
+
+    const serializedProducts = products.map(product => ({
+      product_id: product.id,
+      quantity: product.quantity,
+      price: existentProducts.filter(p => p.id === product.id)[0].price,
+    }));
+
+    const order = await this.ordersRepository.create({
+      customer: customerExist,
+      products: serializedProducts,
+    });
+
+    const { order_products } = order;
+
+    const ordersProductsQuantity = order_products.map(product => ({
+      id: product.product_id,
+      quantity:
+        existentProducts.filter(p => p.id === product.product_id)[0].quantity -
+        product.quantity,
+    }));
+
+    await this.productsRepository.updateQuantity(ordersProductsQuantity);
+    return order;
   }
 }
 
